Added update_todo_item.html and update_todo_item function: - User able to update their todolist items

+++ todolist/views.py
@@ -20,8 +20,24 @@ def complete_todo_item(request, todo_id):
 
 
 def delete_todo_item(request, todo_id):
-	todo_item = TodoItem.objects.filter(id=todo_id).delete()
-	return redirect('todo_list')
+	if request.method == 'DELETE':
+		TodoItem.objects.filter(id=todo_id).delete()
+		return redirect('todo_list')
+
+
+def update_todo_item(request, todo_id):
+	record = TodoItem.objects.get(id=todo_id)
+	if request.method == 'POST':
+		form = TodoItemForm(request.POST, instance=record)
+		if form.is_valid():
+			form.save()
+			return redirect('todo_list')
+		else:
+			form = TodoItemForm(instance=record)
+		return render(request, 'todolist/update_todo_item.html', {'form':form})
+
+	pass
+
 
 def add_todo_item(request):
 	if request.method == 'POST':
